Adjust atmega measurement structure to the others
This adjusts the folder structure in the measurement examples
to include an atmega subfolder, just as we have an nRF and
ESP subfolder.
It also splits the GPIO implementations in the respective
microcontroller parts.
Furthermore, this implements the noop for the atmega HW crypto.
