+++ examples/measurements/atmega/CMakeLists.txt
@@ -1,8 +1,8 @@
-include(../../cmake/linker-map.cmake)
+include(../../../cmake/linker-map.cmake)
 
 function(add_measurement name)
-  add_executable(${name} "common/${name}.c" common/gpio.c)
-  target_include_directories(${name} PRIVATE common/)
+  add_executable(${name} "common/${name}.c" gpio.c hw_crypto.c)
+  target_include_directories(${name} PRIVATE common/ .)
   add_linker_map_for_target(${name})
   target_link_libraries(${name} ${PRODUCT_NAME})
 endfunction()
