+++ examples/measurements/atmega/hw_crypto.c
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2022 Felix Gohla, Konrad Hanff, Tobias Kantusch,
+ *                    Quentin Kuth, Felix Roth. All rights reserved.
+ *
+ * Use of this source code is governed by a BSD-style
+ * license that can be found in the LICENSE file.
+ */
+
+int init_hw_crypto() {
+    return 0;
+}
