+++ examples/measurements/common/aes_gcm_measure.c
@@ -33,9 +33,7 @@ int main(void) {
 
     uint8_t plain[576] = {0};
 
-#ifdef CONFIG_USE_HW_CRYPTO
     init_hw_crypto();
-#endif
 
     setup_pin();
     pin_off();
