+++ examples/measurements/common/sha512_measure.c
@@ -30,9 +30,7 @@ int main(void) {
 
     uint8_t hash[64] = {0};
 
-#ifdef CONFIG_USE_HW_CRYPTO
     init_hw_crypto();
-#endif
 
     setup_pin();
     pin_off();
