+++ examples/measurements/esp32/main/CMakeLists.txt
@@ -1,5 +1,5 @@
 idf_component_register(
-    SRCS "src/${CONFIG_MEASURE_ALGORITHM}_measure.c" "src/gpio.c"
+    SRCS "src/${CONFIG_MEASURE_ALGORITHM}_measure.c" "gpio.c"
     INCLUDE_DIRS "."
     PRIV_REQUIRES libmicrofido2 clock hw_crypto mbedtls driver
 )
