+++ examples/measurements/nrf52/CMakeLists.txt
@@ -8,8 +8,8 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(nrf52-libmicrofido C)
 
-target_sources(app PRIVATE "src/${measure_algorithm}_measure.c" src/gpio.c hw_crypto/hw_crypto.c)
-target_include_directories(app PRIVATE src/ hw_crypto/ clock/)
+target_sources(app PRIVATE "src/${measure_algorithm}_measure.c" gpio/gpio.c hw_crypto/hw_crypto.c)
+target_include_directories(app PRIVATE src/ hw_crypto/ clock/ gpio/)
 
 #######################################
 # libmicrofido2
